Rename libappimage header to appimage.h
